requirements.txt
pandas>=2.2
kafka-python>=2.0
google-cloud-bigquery>=3.10
